[refactor] Separate UT files
